+++ src/main/java/janggi/domain/board/Board.java
@@ -1,6 +1,5 @@
 package janggi.domain.board;
 
-import janggi.domain.Position;
 import janggi.domain.board.initializer.BoardInitializer;
 import janggi.domain.piece.Camp;
 import janggi.domain.piece.Piece;
