+++ src/main/java/janggi/domain/board/initializer/BoardInitializer.java
@@ -1,6 +1,6 @@
 package janggi.domain.board.initializer;
 
-import janggi.domain.Position;
+import janggi.domain.board.Position;
 import janggi.domain.piece.Piece;
 import java.util.Map;
