+++ src/main/java/janggi/domain/piece/strategy/HorseStrategy.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece.strategy;
 
-import janggi.domain.Position;
+import janggi.domain.board.Position;
 import janggi.domain.piece.Camp;
 import java.util.ArrayList;
 import java.util.List;
