+++ src/test/java/janggi/domain/piece/condition/OnePieceExistsConditionTest.java
@@ -2,8 +2,8 @@ package janggi.domain.piece.condition;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.domain.Position;
 import janggi.domain.board.Board;
+import janggi.domain.board.Position;
 import janggi.domain.board.initializer.BoardInitializer;
 import janggi.domain.piece.Camp;
 import janggi.domain.piece.Piece;
